move /gui/wizard.py to /wizard.py
